fix(frontend/trpc/server): pass trpc router with headers
